requirements.txt
python-dotenv==1.0.1
git+https://github.com/ai-cfia/fertiscan-pipeline.git@v0.0.11
git+https://github.com/ai-cfia/ailab-datastore.git@v1.0.17-fertiscan-datastore
psycopg[pool]
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-logging==0.48b0
openinference-instrumentation-dspy==0.1.14
fastapi[standard]==0.115.4
pydantic-settings==2.7.1
pydantic-extra-types==2.10.2
filetype==1.2.0
minio==7.2.15
pillow==10.3.0
